+++ packages/db/src/search/RuleSearch.ts
@@ -3,6 +3,7 @@ import pick from 'lodash/pick';
 
 import Logger from '@cumulus/logger';
 import { RuleRecord } from '@cumulus/types/api/rules';
+
 import { BaseSearch } from './BaseSearch';
 import { DbQueryParameters, QueryEvent } from '../types/search';
 import { PostgresRuleRecord } from '../types/rule';
@@ -26,78 +27,72 @@ export class RuleSearch extends BaseSearch {
   }
 
   /**
-  * Build basic query
-  *
-  * @param knex - DB client
-  * @returns queries for getting count and search result
-  */
-  protected buildBasicQuery(knex: Knex): {
+   * Build basic query
+   *
+   * @param knex - DB client
+   * @returns count query and joined CTE search query
+   */
+  protected buildBasicQuery(knex: Knex) : {
     countQuery: Knex.QueryBuilder,
-    searchQuery: Knex.QueryBuilder,
+    cteQueryBuilder: Knex.QueryBuilder,
   } {
     const {
       collections: collectionsTable,
       providers: providersTable,
     } = TableNames;
 
-    const countQuery = knex(this.tableName)
-      .count(`${this.tableName}.cumulus_id`);
+    const cteQueryBuilder = knex(this.tableName)
+      .select(
+        `${this.tableName}.*`,
+        `${collectionsTable}.name as collectionName`,
+        `${collectionsTable}.version as collectionVersion`,
+        `${providersTable}.name as providerName`
+      )
+      .leftJoin(collectionsTable, `${this.tableName}.collection_cumulus_id`, `${collectionsTable}.cumulus_id`)
+      .leftJoin(providersTable, `${this.tableName}.provider_cumulus_id`, `${providersTable}.cumulus_id`);
 
-    const searchQuery = knex(this.tableName)
-      .select(`${this.tableName}.*`)
-      .select({
-        collectionName: `${collectionsTable}.name`,
-        collectionVersion: `${collectionsTable}.version`,
-        providerName: `${providersTable}.name`,
-      });
+    const countQuery = knex(this.tableName).count(`${this.tableName}.cumulus_id`);
 
     if (this.searchCollection()) {
-      searchQuery.innerJoin(collectionsTable, `${this.tableName}.collection_cumulus_id`, `${collectionsTable}.cumulus_id`);
       countQuery.innerJoin(collectionsTable, `${this.tableName}.collection_cumulus_id`, `${collectionsTable}.cumulus_id`);
-    } else {
-      searchQuery.leftJoin(collectionsTable, `${this.tableName}.collection_cumulus_id`, `${collectionsTable}.cumulus_id`);
     }
 
     if (this.searchProvider()) {
-      searchQuery.innerJoin(providersTable, `${this.tableName}.provider_cumulus_id`, `${providersTable}.cumulus_id`);
       countQuery.innerJoin(providersTable, `${this.tableName}.provider_cumulus_id`, `${providersTable}.cumulus_id`);
-    } else {
-      searchQuery.leftJoin(providersTable, `${this.tableName}.provider_cumulus_id`, `${providersTable}.cumulus_id`);
     }
 
-    return { countQuery, searchQuery };
+    return { countQuery, cteQueryBuilder };
   }
 
   /**
-  * Build queries for infix and prefix
-  *
-  * @param params
-  * @param params.countQuery - query builder for getting count
-  * @param params.searchQuery - query builder for search
-  * @param [params.dbQueryParameters] - db query parameters
-  */
+   * Build queries for infix and prefix
+   *
+   * @param params
+   * @param params.countQuery - knex query for count
+   * @param params.cteQueryBuilder - CTE query builder
+   * @param [params.dbQueryParameters] - db query parameters
+   */
   protected buildInfixPrefixQuery(params: {
     countQuery: Knex.QueryBuilder,
-    searchQuery: Knex.QueryBuilder,
+    cteQueryBuilder: Knex.QueryBuilder,
     dbQueryParameters?: DbQueryParameters,
   }) {
-    const { countQuery, searchQuery, dbQueryParameters } = params;
+    const { countQuery, cteQueryBuilder, dbQueryParameters } = params;
     const { infix, prefix } = dbQueryParameters ?? this.dbQueryParameters;
     if (infix) {
-      [countQuery, searchQuery].forEach((query) => query.whereLike(`${this.tableName}.name`, `%${infix}%`));
+      [countQuery, cteQueryBuilder].forEach((query) => query.whereLike(`${this.tableName}.name`, `%${infix}%`));
     }
     if (prefix) {
-      [countQuery, searchQuery].forEach((query) => query.whereLike(`${this.tableName}.name`, `${prefix}%`));
+      [countQuery, cteQueryBuilder].forEach((query) => query.whereLike(`${this.tableName}.name`, `${prefix}%`));
     }
   }
 
   /**
-  * Translate postgres records to api records
-  *
-  * @param pgRecords - postgres Rule records returned from query
-  * @param knex - knex for the translation method
-  * @returns translated api records
-  */
+   * Translate postgres records to api records
+   *
+   * @param pgRecords - postgres Rule records returned from query
+   * @returns translated api records
+   */
   protected async translatePostgresRecordsToApiRecords(
     pgRecords: RuleRecordWithExternals[]
   ): Promise<Partial<RuleRecord>[]> {