+++ CHANGELOG.md
@@ -14,6 +14,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ### Fixed
 
+- **CUMULUS-4090**
+  - add limit = number of granules to get return out as requested from listGranules
+
+### Changed
+
+- **CUMULUS-4096**
+  - optimized BaseSearch and child class queries to use CTE (common table expression) for improved performance
 - **CUMULUS-4106**
   - Fixed the release npm publish error by adding private property to `@cumulus/change-granule-collection-pg`
     and `@cumulus/change-granule-collection-s3` package.json.
@@ -51,8 +58,6 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ### Fixed
 
-- **CUMULUS-4090**
-  - add limit = number of granules to get return out as requested from listGranules
 - **CUMULUS-3868**
   - exclude package cloudflare:sockets" in webpack.config throughout to prevent packaging bug
 - **CUMULUS-3752**
@@ -67,7 +72,6 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 - **CUMULUS-3868**
   - Use listGranules endpoint to gather granules from granuleIds
   - Bound cmr metadata download in post-to-cmr task to s3Concurrency limit
-
 - **CUMULUS-3788**
   - Updated `@cumulus/launchpad-auth/getLaunchpadToken` to check if the token in s3 has been updated
     before updating it with a new token